browser-use/pyproject.toml
[project]
name = "browser-use"
description = "Make websites accessible for AI agents"
authors = [{ name = "Link" }]
version = "0.12.5"
readme = "README.md"
requires-python = ">=3.11,<4.0"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: Other/Proprietary License",
    "Operating System :: OS Independent",
]
dependencies = [
    "aiohttp==3.13.3",
    "anyio==4.12.1",
    "bubus==1.5.6",
    "click==8.3.1",
    "InquirerPy==0.3.4",
    "rich==14.3.1",
    "google-api-core==2.29.0",
    "httpx==0.28.1",
    "posthog==7.7.0",
    "psutil==7.2.2",
    "pydantic==2.12.5",
    "pyobjc==12.1; platform_system == 'darwin'",
    "python-dotenv==1.2.1",
    "requests==2.32.5",
    "screeninfo==0.8.1; platform_system != 'darwin'",
    "typing-extensions==4.15.0",
    "uuid7==0.1.0",
    "google-genai==1.65.0",
    "openai==2.16.0",
    "anthropic==0.76.0",
    "groq==1.0.0",
    "ollama==0.6.1",
    "google-api-python-client==2.188.0",
    "google-auth==2.48.0",
    "google-auth-oauthlib==1.2.4",
    "mcp==1.26.0",
    "pypdf==6.9.1",
    "reportlab==4.4.9",
    "cdp-use==1.4.5",
    "pyotp==2.9.0",
    "pillow==12.1.1",
    "cloudpickle==3.1.2",
    "markdownify==1.2.2",
    "python-docx==1.2.0",
    "browser-use-sdk==2.0.15",
]
# google-api-core: only used for Google LLM APIs
# pyperclip: only used for examples that use copy/paste
# pyobjc: only used to get screen resolution on macOS
# screeninfo: only used to get screen resolution on Linux/Windows
# markdownify: used for page text content extraction for passing to LLM
# openai: datalib,voice-helpers are actually NOT NEEDED but openai produces noisy errors on exit without them TODO: fix
# rich: used for terminal formatting and styling in CLI
# click: used for command-line argument parsing
# textual: used for terminal UI

[project.optional-dependencies]
cli = ["textual==7.4.0"]
ocr = ["paddleocr>=2.7", "paddlepaddle>=2.6"]
aws = ["boto3==1.42.37"]
oci = ["oci==2.166.0"]
video = ["imageio[ffmpeg]==2.37.2", "numpy==2.4.1"]
examples = [
    "agentmail==0.0.59",
    # botocore: only needed for Bedrock Claude boto3 examples/models/bedrock_claude.py
    "botocore==1.42.37",
    "imgcat==0.6.0",
    # "stagehand-py>=0.3.6",
    # "browserbase>=0.4.0",
    "langchain-openai==1.1.7",
]
eval = [
    "lmnr[all]==0.7.42",
    "anyio==4.12.1",
    "psutil==7.2.2",
    "datamodel-code-generator==0.53.0",
]
cli-oci = ["browser-use[cli,oci]"]
all = ["browser-use[cli,examples,aws,oci]"]

# will prefer to use local source code checked out in ../../browser-use (if present) instead of pypi browser-use package
# [tool.uv.sources]
# bubus = { path = "../bubus", editable = true }


[project.urls]
Homepage = "https://browser-use.com"
Documentation = "https://docs.browser-use.com"
Repository = ""
Telemetry = "https://docs.browser-use.com/development/monitoring/telemetry"
"Terms of Service" = "https://browser-use.com/legal/terms-of-service"
"Privacy Policy" = "https://browser-use.com/privacy/"

[project.scripts]
browser-use = "browser_use.skill_cli.main:main"  # Fast CLI for browser automation
browseruse = "browser_use.skill_cli.main:main"  # Alias for browser-use
bu = "browser_use.skill_cli.main:main"  # Alias for browser-use
browser = "browser_use.skill_cli.main:main"  # Alias for browser-use
browser-use-tui = "browser_use.cli:main"  # Legacy TUI interface

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"


[tool.codespell]
ignore-words-list = "bu,wit,dont,cant,wont,re-use,re-used,re-using,re-usable,thats,doesnt,doubleclick,finaly,finalY"
skip = "*.json"

[tool.ruff]
line-length = 130
fix = true

[tool.ruff.lint]
select = ["ASYNC", "E", "F", "FAST", "I", "PLE"]
ignore = [
    "ASYNC109",
    "E101",
    "E402",
    "E501",
    "F841",
    "E731",
    "W291",
] # TODO: determine if adding timeouts to all the unbounded async functions is needed / worth-it so we can un-ignore ASYNC109
unfixable = ["E101", "E402", "E501", "F841", "E731"]

[tool.ruff.format]
quote-style = "single"
indent-style = "tab"
line-ending = "lf"
docstring-code-format = true
docstring-code-line-length = 140
skip-magic-trailing-comma = false

[tool.pyright]
typeCheckingMode = "basic"
include = ["browser_use", "examples", "tests"]
exclude = [
    ".venv/",
    ".venv*/",
    ".git/",
    "__pycache__/",
    "**/site-packages/",
    "./test_*.py",
    "./debug_*.py",
    "private_example/",
    "debug/*",
    "tests/scripts/*",
    "tests/old/*",
    "browser_use/dom/playground/*",
    "examples/use-cases/onepassword.py",
    "browser_use/llm/oci_raw/*",
    "browser_use/llm/tests/test_chat_models.py",
    "browser_use/llm/tests/test_single_step.py",
    "product_extraction.py",
    "discover/",
    "list/",
]
venvPath = "."
venv = ".venv"
reportMissingTypeStubs = false


[tool.hatch.build]
include = [
    "browser_use/**/*.py",
    "!browser_use/**/tests/*.py",
    "!browser_use/**/tests.py",
	"browser_use/agent/system_prompts/*.md",
    "browser_use/cli_templates/*.py",
    "browser_use/py.typed",
    "browser_use/dom/**/*.js",
    "!tests/**/*.py",
    "!debug/*",
]

[tool.pytest.ini_options]
timeout = 300
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
asyncio_default_test_loop_scope = "session"
markers = [
    "slow: marks tests as slow (deselect with `-m 'not slow'`)",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
    "asyncio: mark tests as async tests",
]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
addopts = "-svx --strict-markers --tb=short --dist=loadscope"
log_cli = true
log_cli_format = "%(levelname)-8s [%(name)s] %(message)s"
filterwarnings = [
    "ignore::pytest.PytestDeprecationWarning",
    "ignore::DeprecationWarning",
]
log_level = "DEBUG"


[tool.hatch.metadata]
allow-direct-references = true

[tool.uv]
# required-environments = [
#     "sys_platform == 'darwin' and platform_machine == 'arm64'",
#     "sys_platform == 'darwin' and platform_machine == 'x86_64'",
#     "sys_platform == 'linux' and platform_machine == 'x86_64'",
#     "sys_platform == 'linux' and platform_machine == 'aarch64'",
#     # "sys_platform == 'linux' and platform_machine == 'arm64'",  # no pytorch wheels available yet
#     "sys_platform == 'win32' and platform_machine == 'x86_64'",
#     # "sys_platform == 'win32' and platform_machine == 'arm64'",  # no pytorch wheels available yet
# ]
dev-dependencies = [
    "ruff==0.14.14",
    "tokencost==0.1.26",
    "build==1.4.0",
    "pytest==9.0.2",
    "pytest-asyncio==1.3.0",
    "pytest-httpserver==1.1.3",
    "fastapi==0.128.0",
    "inngest==0.5.15",
    "uvicorn==0.40.0",
    "ipdb==0.13.13",
    "pre-commit==4.5.1",
    "codespell==2.4.1",
    "pyright==1.1.408",
    "ty==0.0.14",
    "pytest-xdist==3.8.0",
    "lmnr[all]==0.7.42",
    # "pytest-playwright-asyncio>=0.7.0",  # not actually needed I think
    "pytest-timeout==2.4.0",
    "pydantic_settings==2.12.0",
]
